clean up events, made events look good when 1 or 2 events are showing, also moved code segments around
