+++ lantern-club/src/components/events/EventsII.tsx
@@ -4,6 +4,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTrashCan } from "@fortawesome/free-solid-svg-icons";
 import { EventType } from "@/types/event";
 import ConfirmModal from "../ConfirmModal";
+import { EventBox } from "./EventBox";
 
 type EventsIIProps = {
   isAdminEdit: boolean;
@@ -15,7 +16,6 @@ type EventsIIProps = {
 export default function EventsII({
   events,
   isAdminEdit,
-  handleEditButtonClick,
 }: EventsIIProps) {
   const header_font: React.CSSProperties = {
     fontFamily: "coolvetica",
@@ -30,31 +30,6 @@ export default function EventsII({
   const imageStyle: React.CSSProperties = {
     width: "100%",
     height: "100%",
-  };
-
-  const [showConfirmModal, setShowConfirmModal] = useState(false);
-
-  const handleDelete = async (id: string) => {
-    const url = `/api/content/events/${id}`;
-    try {
-      const response = await fetch(url, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-
-      const deletedEvent = await response.json();
-      console.log("Deleted profile:", deletedEvent);
-      window.location.reload();
-      setShowConfirmModal(false);
-    } catch (error) {
-      console.error("Failed to delete event:", error);
-    }
   };
 
   
@@ -73,39 +48,11 @@ export default function EventsII({
               .slice(0, 3)
               .map((oneEvent: EventType) => {
                 return (
-                  <div
-                    key={oneEvent.id}
-                    className="flex-1"
-                    style={{
-                      display: "flex",
-                      flexDirection: "column",
-                      alignItems: "center",
-                    }}
-                  >
-                    <div className="relative flex-shrink w-full" style={{ minHeight: '500px'}}>
-                              <Image
-                                  src={oneEvent.imageURL}
-                                  style={imageStyle}
-                                  layout="fill"
-                                  objectFit="cover"
-                                  alt="Event image"
-                                  className="rounded-xl overflow-hidden"
-                              />
-                    </div>       
-                    {isAdminEdit && (
-                      <div style={{ paddingTop: "20px" }}>
-                        <FontAwesomeIcon
-                          icon={faTrashCan}
-                          width={100}
-                          height={100}
-                          size="2x"
-                          onClick={() => setShowConfirmModal(true)}
-                          className="cursor-pointer relative transition-all duration-300 hover:text-orange-400"
-                        />
-                        <ConfirmModal isVisible={showConfirmModal} onClose={() => {setShowConfirmModal(false)}} onDelete={() => oneEvent.id && handleDelete(oneEvent.id)} />
-                      </div>
-                    )}
+                  <div key={oneEvent.id} className="w-full md:w-1/3">
+                    <EventBox event={oneEvent} isAdminEdit={isAdminEdit} isPast={true}/>
                   </div>
+              
+                
                 );
               })}
         </div>
